Web/Blog/blog/requirements/requirements.txt
Django==3.0.3
django-extensions==2.2.8
djangorestframework==3.11.0
